+++ R/utilities.R
@@ -576,8 +576,9 @@ p
 
 # Summary functions
 .summary_functions <- function(){
-  c("mean", "mean_se", "mean_sd", "mean_ci",
-    "mean_range", "median", "median_iqr", "median_mad", "median_range")
+  c("mean", "mean_se", "mean_se_", "mean_sd", "mean_ci",
+    "mean_range", "median", "median_iqr", "median_hilow", "median_hilow_",
+    "median_q1q3", "median_mad", "median_range")
 }
 .errorbar_functions <- function(){
   setdiff(.summary_functions(), c("mean", "median"))